Migration changes, Id type changed to long, brands and locations

+++ CustomerGreen.Core/Entities/CheckoutPoint.cs
@@ -2,7 +2,7 @@
 {
     public class CheckoutPoint : BaseEntity
     {
-        public int OrgId { get; set; }
+        public long OrgId { get; set; }
         public string CheckoutLocation { get; set; }
     }
 }
