+++ CustomerGreen.Core/Entities/Question.cs
@@ -2,7 +2,7 @@
 {
     public class Question : BaseEntity
     {
-        public int OrgId { get; set; }
+        public long OrgId { get; set; }
         public string CategoryKey { get; set; }
         public string FeedbackQuestion { get; set; }
     }
